requirements.txt
fastapi
uvicorn
python-dotenv
langchain-core
langchain-openai
requests
pydantic
openai
fastapi-cors
fastapi-websocket-rpc
flask